packages/qna-poc/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
chromadb>=0.5.0
networkx>=3.2
gradio>=4.20.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.30.0
